Each proxy carries both formatting hooks, and they agree by construction (#131)
The std::formatter specializations move out of xstd/bits/format.hpp and into
the proxy headers that define the types they format, and format.hpp goes
away.

They also stop restating the value. The formatter used to reach it through
the conversion operator -- static_cast<size_t>(ref), static_cast<bool>(ref)
-- while format_as read the member. Both landed on the same value and
nothing made them: changing one would have left fmt::format and std::format
disagreeing about the same object, with no compile error and no test to
catch it, since the proxy tests check format_as and the format test checks
std::format but nothing checked that the two agree. The formatter now calls
format_as, unqualified so ADL finds the proxy's own hidden friend, and
format_as is the one place that says what a proxy prints as.

Why the separate header is gone. It was outside the umbrella to keep
<format> off the path of a consumer who does not format, and that reason had
quietly expired: sequence_adaptor and bitset_adaptor already include
<format> to build the file:line:column: messages their exceptions carry.
Measured per column, the separate header was saving the sequence reading 181
preprocessed lines and the bitset reading 379. Only the set reading paid
anything real, about 11.8k, set_adaptor being the one adaptor that does not
format its own errors.

Against that: hashing and formatting are both customization points a
user-defined type owes the standard library, and std::hash has always been
specialized inside the adaptor headers. One type's two protocols now live in
one file, and the two customization points are consistent with each other.
The standard cannot do this for its own types -- formatter<pair> lives in
<format> because <utility> must not depend on it -- but that arrow only
points that way for types the standard owns; <format> will never know about
a proxy of ours.

design.md records the direction #20 argued to WG21 on P3070R0: one format_as
in the type's own namespace could serve both libraries, where a formatter
specialization serves one. Neither libc++ nor the MSVC STL defines such a
hook today, so the specialization stays -- it just no longer restates the
value. And it records the other half of that argument, which this library is
the worked example of: a proxy nested in its container cannot be named by
either hook, which is why the references are classes of their own.

README stops telling users to include a header that no longer exists;
formatting now needs nothing included and nothing switched on. The mirrored
test moves to test/src/bits/detail/format.cpp and keeps every case.

Verified: 32/32 headers self-sufficient; every test source compiles whole
(16 clean, 26 skipped for deps this container lacks, 0 failures); both
proxies format with no formatting header in the TU, the whole spec intact
({:#x}, {:*>6}, {:d}); and format_as and std::format agree on both.
