+++ include/xstd/bits/detail/random_access.hpp
@@ -11,6 +11,7 @@
 #include <compare>                  // strong_ordering
 #include <concepts>                 // same_as
 #include <cstddef>                  // ptrdiff_t, size_t
+#include <format>                   // formatter
 #include <iterator>                 // random_access_iterator_tag
 #include <type_traits>              // is_class_v, is_const_v, is_convertible_v, is_nothrow_constructible_v, remove_const_t
 
@@ -196,5 +197,35 @@ public:
 };
 
 }       // namespace xstd::detail::bits
+
+
+// std::format over the containers, which needs nothing said about the containers themselves.
+// [design.md#formatting-the-proxies]
+//
+// Every owner and view here is already a range, so [format.range.formatter] would format it -- except that the
+// range formatter requires formattable<range_reference_t<R>>, and a reference of ours is a proxy. So the proxy
+// is what gets a formatter, and every container over it follows.
+//
+// It defers to format_as, the hook fmt already calls, so the value this proxy prints as is defined once and both
+// libraries read it from there. Deriving from the underlying formatter rather than writing parse() is what keeps
+// the whole format spec: a width, a fill, {:#x} on a position and {:d} on a bool, and the nested spec a range
+// formatter forwards ({::#x}) reaching them.
+//
+// [namespace.std]/2 allows a specialization of a standard library template for a program-defined type, which is
+// what this is and all it is. clang-tidy 22 and 23 read the qualified definition as modifying namespace std
+// anyway; 24 no longer does. [design.md#clang-tidy-false-positives]
+template<class Bits, class Traits, class CharT>
+// NOLINTNEXTLINE(bugprone-std-namespace-modification)
+struct std::formatter<xstd::detail::bits::random_access_bit_reference<Bits, Traits>, CharT>
+:
+        std::formatter<bool, CharT>
+{
+        template<class Context>
+        [[nodiscard]] constexpr auto format(xstd::detail::bits::random_access_bit_reference<Bits, Traits> ref, Context& ctx) const
+        {
+                // Unqualified, so ADL finds the proxy's own hidden friend. [design.md#the-one-adl-exception]
+                return std::formatter<bool, CharT>::format(format_as(ref), ctx);
+        }
+};
 
 #endif  // XSTD_BITS_DETAIL_RANDOM_ACCESS_HPP
